requirements.txt
flask==3.0.0
google-generativeai==0.3.1
Werkzeug==3.0.1
gunicorn==21.2.0
